# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Oralia API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [profile]
      summary: Create or update user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileBody"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
...
  /relationships:
    get:
      operationId: listRelationships
      tags: [relationships]
      summary: List relationship profiles
      responses:
        "200":
          description: List of relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RelationshipProfile"
    post:
      operationId: createRelationship
      tags: [relationships]
      summary: Create a relationship profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRelationshipBody"
      responses:
        "201":
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipProfile"
  /relationships/{id}:
    get:
      operationId: getRelationship
      tags: [relationships]
      summary: Get a relationship profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Relationship profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipProfile"
        "404":
          description: Relationship not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateRelationship
      tags: [relationships]
      summary: Update a relationship profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRelationshipBody"
      responses:
        "200":
          description: Relationship updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipProfile"
        "404":
          description: Relationship not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRelationship
      tags: [relationships]
      summary: Delete a relationship profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Relationship deleted
  /relationships/{id}/summary:
    post:
      operationId: generateRelationshipSummary
      tags: [relationships]
      summary: Generate AI relationship summary
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Relationship summary generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RelationshipProfile"
        "404":
          description: Relationship not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Profile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        birthday:
          type: string
        birthTime:
          type: ["string", "null"]
        birthCity:
          type: ["string", "null"]
        currentCity:
          type: ["string", "null"]
        relationshipStatus:
          type: ["string", "null"]
        careerStage:
          type: ["string", "null"]
        topGoals:
          type: array
          items:
            type: string
        currentChallenges:
          type: ["string", "null"]
        guidanceCategories:
          type: array
          items:
            type: string
        menstrualCycleTracking:
          type: boolean
        sleepTracking:
          type: boolean
        spiritualOpenness:
          type: ["string", "null"]
        guidanceTone:
          type: ["string", "null"]
        hdType:
          type: ["string", "null"]
        hdStrategy:
          type: ["string", "null"]
        hdAuthority:
          type: ["string", "null"]
        hdProfile:
          type: ["string", "null"]
        hdDefinedCenters:
          type: array
          items:
            type: string
        hdKeyGates:
          type: array
          items:
            type: string
        sunSign:
          type: ["string", "null"]
        lifePathNumber:
          type: ["integer", "null"]
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - fullName
        - birthday
        - onboardingComplete
    UpdateProfileBody:
      type: object
      properties:
        fullName:
          type: string
        birthday:
          type: string
        birthTime:
          type: ["string", "null"]
        birthCity:
          type: ["string", "null"]
        currentCity:
          type: ["string", "null"]
        relationshipStatus:
          type: ["string", "null"]
        careerStage:
          type: ["string", "null"]
        topGoals:
          type: array
          items:
            type: string
        currentChallenges:
          type: ["string", "null"]
        guidanceCategories:
          type: array
          items:
            type: string
        menstrualCycleTracking:
          type: boolean
        sleepTracking:
          type: boolean
        spiritualOpenness:
          type: ["string", "null"]
        guidanceTone:
          type: ["string", "null"]
        hdType:
          type: ["string", "null"]
        hdStrategy:
          type: ["string", "null"]
        hdAuthority:
          type: ["string", "null"]
        hdProfile:
          type: ["string", "null"]
        hdDefinedCenters:
          type: array
          items:
            type: string
        hdKeyGates:
          type: array
          items:
            type: string
        onboardingComplete:
          type: boolean
      required:
        - fullName
        - birthday
...
    RelationshipProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        personName:
          type: string
        relationshipType:
          type: string
        birthday:
          type: ["string", "null"]
        birthTime:
          type: ["string", "null"]
        birthCity:
          type: ["string", "null"]
        communicationStyle:
          type: ["string", "null"]
        attachmentStyle:
          type: ["string", "null"]
        conflictStyle:
          type: ["string", "null"]
        loveLanguage:
          type: ["string", "null"]
        currentDynamic:
          type: ["string", "null"]
        communicationPattern:
          type: ["string", "null"]
        emotionalActivation:
          type: ["string", "null"]
        repairLanguage:
          type: ["string", "null"]
        conflictPattern:
          type: ["string", "null"]
        greenFlags:
          type: ["string", "null"]
        redFlags:
          type: ["string", "null"]
        bestCommunication:
          type: ["string", "null"]
        bestTiming:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - personName
        - relationshipType
    CreateRelationshipBody:
      type: object
      properties:
        personName:
          type: string
        relationshipType:
          type: string
        birthday:
          type: ["string", "null"]
        birthTime:
          type: ["string", "null"]
        birthCity:
          type: ["string", "null"]
        communicationStyle:
          type: ["string", "null"]
        attachmentStyle:
          type: ["string", "null"]
        conflictStyle:
          type: ["string", "null"]
        loveLanguage:
          type: ["string", "null"]
        currentDynamic:
          type: ["string", "null"]
      required:
        - personName
        - relationshipType
    UpdateRelationshipBody:
      type: object
      properties:
        personName:
          type: string
        relationshipType:
          type: string
        birthday:
          type: ["string", "null"]
        birthTime:
          type: ["string", "null"]
        birthCity:
          type: ["string", "null"]
        communicationStyle:
          type: ["string", "null"]
        attachmentStyle:
          type: ["string", "null"]
        conflictStyle:
          type: ["string", "null"]
        loveLanguage:
          type: ["string", "null"]
        currentDynamic:
          type: ["string", "null"]
...