# lib/api-zod/src/generated/types/profile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Oralia API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Profile {
  id: number;
  userId: number;
  fullName: string;
  birthday: string;
  /** @nullable */
  birthTime?: string | null;
  /** @nullable */
  birthCity?: string | null;
  /** @nullable */
  currentCity?: string | null;
  /** @nullable */
  relationshipStatus?: string | null;
  /** @nullable */
  careerStage?: string | null;
  topGoals?: string[];
  /** @nullable */
  currentChallenges?: string | null;
  guidanceCategories?: string[];
  menstrualCycleTracking?: boolean;
  sleepTracking?: boolean;
  /** @nullable */
  spiritualOpenness?: string | null;
  /** @nullable */
  guidanceTone?: string | null;
  /** @nullable */
  hdType?: string | null;
  /** @nullable */
  hdStrategy?: string | null;
  /** @nullable */
  hdAuthority?: string | null;
  /** @nullable */
  hdProfile?: string | null;
  hdDefinedCenters?: string[];
  hdKeyGates?: string[];
  /** @nullable */
  sunSign?: string | null;
  /** @nullable */
  lifePathNumber?: number | null;
  onboardingComplete: boolean;
  createdAt?: Date;
  updatedAt?: Date;
}